Factor out application job-related request method tests
